# specs/001-cicd-training-app/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: CI/CD Training Application API
  description: |
    Backend API for the CI/CD Training Application demonstrating deployment automation patterns.
    This API provides health checks, metrics, and deployment management endpoints for training purposes.
  version: 1.0.0
  contact:
    name: DevOps Training Team
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://alpha.example.com
    description: Alpha environment
  - url: https://beta.example.com
    description: Beta environment
  - url: https://nonprod.example.com
    description: Non-production environment
  - url: https://prod.example.com
    description: Production environment
...
paths:
...
  /api/v1/environments/{id}:
    get:
      tags:
        - Environments
      summary: Get environment by ID
      operationId: getEnvironment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Environment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Environment'
        '404':
          description: Environment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/releases:
    get:
      tags:
        - Releases
      summary: List all releases
      description: Returns paginated list of software releases
      operationId: listReleases
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: List of releases
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Release'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Releases
      summary: Create a new release
      description: Creates a new software release with version and artifacts
      operationId: createRelease
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReleaseRequest'
      responses:
        '201':
          description: Release created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Release'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Release version already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/releases/{id}:
    get:
      tags:
        - Releases
      summary: Get release by ID
      operationId: getRelease
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Release details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Release'
        '404':
          description: Release not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/deployments:
    get:
      tags:
        - Deployments
      summary: List deployments
      description: Returns paginated list of deployments with optional filters
      operationId: listDeployments
      parameters:
        - name: environment_id
          in: query
          schema:
            type: string
            format: uuid
        - name: release_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/DeploymentStatus'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of deployments
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Deployment'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Deployments
      summary: Create a new deployment
      description: Initiates a deployment of a release to an environment
      operationId: createDeployment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeploymentRequest'
      responses:
        '201':
          description: Deployment initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Deployment already in progress for environment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/deployments/{id}:
    get:
      tags:
        - Deployments
      summary: Get deployment by ID
      operationId: getDeployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Deployment details with health checks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentDetails'
        '404':
          description: Deployment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/deployments/{id}/rollback:
    post:
      tags:
        - Deployments
      summary: Rollback a deployment
      description: Creates a rollback deployment to the previous successful version
      operationId: rollbackDeployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                triggered_by:
                  type: string
                  description: User or system initiating rollback
              required:
                - triggered_by
      responses:
        '201':
          description: Rollback deployment initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
        '400':
          description: Cannot rollback this deployment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    Environment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: alpha
        url:
          type: string
          format: uri
          example: https://alpha.example.com
        approval_required:
          type: boolean
        is_production:
          type: boolean
        soak_period_hours:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - url
    Release:
      type: object
      properties:
        id:
          type: string
          format: uuid
        version:
          type: string
          pattern: '^\d+\.\d+\.\d+(-[a-zA-Z0-9.]+)?$'
          example: "1.2.3"
        git_sha:
          type: string
          minLength: 40
          maxLength: 40
          example: "abc123def456789012345678901234567890abcd"
        git_sha_short:
          type: string
          minLength: 7
          maxLength: 7
          example: "abc123d"
        changelog:
          type: string
          nullable: true
        docker_image_backend:
          type: string
          example: "registry.example.com/backend:1.2.3"
        docker_image_frontend:
          type: string
          example: "registry.example.com/frontend:1.2.3"
        created_at:
          type: string
          format: date-time
        created_by:
          type: string
      required:
        - id
        - version
        - git_sha
        - git_sha_short
        - docker_image_backend
        - docker_image_frontend
        - created_by
    CreateReleaseRequest:
      type: object
      properties:
        version:
          type: string
          pattern: '^\d+\.\d+\.\d+(-[a-zA-Z0-9.]+)?$'
        git_sha:
          type: string
          minLength: 40
          maxLength: 40
        changelog:
          type: string
          nullable: true
        docker_image_backend:
          type: string
        docker_image_frontend:
          type: string
        created_by:
          type: string
      required:
        - version
        - git_sha
        - docker_image_backend
        - docker_image_frontend
        - created_by
    DeploymentStatus:
      type: string
      enum:
        - pending
        - running
        - success
        - failed
        - rolled_back
    TriggerType:
      type: string
      enum:
        - manual
        - automatic
        - hotfix
        - rollback
    Deployment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        environment_id:
          type: string
          format: uuid
        release_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/DeploymentStatus'
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        triggered_by:
          type: string
        trigger_type:
          $ref: '#/components/schemas/TriggerType'
        rollback_of_id:
          type: string
          format: uuid
          nullable: true
        error_message:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
      required:
        - id
        - environment_id
        - release_id
        - status
        - triggered_by
        - trigger_type
    DeploymentDetails:
      allOf:
        - $ref: '#/components/schemas/Deployment'
        - type: object
          properties:
            environment:
              $ref: '#/components/schemas/Environment'
            release:
              $ref: '#/components/schemas/Release'
            health_checks:
              type: array
              items:
                $ref: '#/components/schemas/HealthCheck'
    CreateDeploymentRequest:
      type: object
      properties:
        environment_id:
          type: string
          format: uuid
        release_id:
          type: string
          format: uuid
        triggered_by:
          type: string
        trigger_type:
          $ref: '#/components/schemas/TriggerType'
      required:
        - environment_id
        - release_id
        - triggered_by
        - trigger_type
    HealthCheck:
      type: object
      properties:
        id:
          type: string
          format: uuid
        deployment_id:
          type: string
          format: uuid
        check_type:
          type: string
          enum:
            - liveness
            - readiness
        status:
          type: string
          enum:
            - passing
            - failing
        response_time_ms:
          type: integer
          nullable: true
        status_code:
          type: integer
          nullable: true
        error_message:
          type: string
          nullable: true
        checked_at:
          type: string
          format: date-time
      required:
        - id
        - deployment_id
        - check_type
        - status
        - checked_at
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
          required:
            - code
            - message
      required:
        - error